+++ java/com/turbovnc/vncviewer/Viewport.java
@@ -194,6 +194,8 @@ public class Viewport extends JFrame {
         }
       }
     });
+
+    lastEvent.deviceID = -1;
   }
 
   public Dimension getAvailableSize() {
@@ -442,6 +444,79 @@ public class Viewport extends JFrame {
         vlog.info("  " + e.toString());
         vlog.info("  Extended input device support may not work correctly.");
       }
+      if (VncViewer.os.startsWith("mac os x")) {
+        // Create default devices for Wacom tablet
+        for (int i = 0; i < 2; i++) {
+          ExtInputDevice dev = new ExtInputDevice();
+          dev.name = new String(i == 0 ? "Stylus" : "Eraser");
+          dev.id = i;
+          dev.vendorID = 4242;
+          dev.productID = (i == 0 ? giiTypes.giiDevTypeStylus :
+                                    giiTypes.giiDevTypeEraser);
+          dev.canGenerate = giiTypes.giiButtonPressMask |
+                            giiTypes.giiButtonReleaseMask |
+                            giiTypes.giiValuatorAbsoluteMask;
+          dev.numButtons = 3;
+          dev.absolute = true;
+
+          ExtInputDevice.Valuator val = dev.new Valuator();
+          val.index = 0;
+          val.longName = new String("Abs X");
+          val.shortName = new String("0");
+          val.rangeMin = 0;
+          val.rangeMax = 31496;
+          val.rangeCenter = 15748;
+          val.siUnit = giiTypes.giiUnitLength;
+          val.siDiv = 200000;
+          dev.addValuator(val);
+
+          val = dev.new Valuator();
+          val.index = 1;
+          val.longName = new String("Abs Y");
+          val.shortName = new String("1");
+          val.rangeMin = 0;
+          val.rangeMax = 19685;
+          val.rangeCenter = 9843;
+          val.siUnit = giiTypes.giiUnitLength;
+          val.siDiv = 200000;
+          dev.addValuator(val);
+
+          val = dev.new Valuator();
+          val.index = 2;
+          val.longName = new String("Abs Pressure");
+          val.shortName = new String("2");
+          val.rangeMin = 0;
+          val.rangeMax = 65536;
+          val.rangeCenter = 32768;
+          val.siUnit = giiTypes.giiUnitLength;
+          val.siDiv = 1;
+          dev.addValuator(val);
+
+          val = dev.new Valuator();
+          val.index = 3;
+          val.longName = new String("Abs Tilt X");
+          val.shortName = new String("3");
+          val.rangeMin = -64;
+          val.rangeMax = 63;
+          val.rangeCenter = 0;
+          val.siUnit = giiTypes.giiUnitLength;
+          val.siDiv = 57;
+          dev.addValuator(val);
+
+          val = dev.new Valuator();
+          val.index = 4;
+          val.longName = new String("Abs Tilt Y");
+          val.shortName = new String("4");
+          val.rangeMin = -64;
+          val.rangeMax = 63;
+          val.rangeCenter = 0;
+          val.siUnit = giiTypes.giiUnitLength;
+          val.siDiv = 57;
+          dev.addValuator(val);
+
+          addInputDevice(dev);
+        }
+      }
     }
   }
 
@@ -487,7 +562,8 @@ public class Viewport extends JFrame {
 
   boolean processExtInputEventHelper(int type) {
     boolean retval = false;
-    if (isHelperAvailable() && cc.cp.supportsGII) {
+    if (isHelperAvailable() && cc.cp.supportsGII &&
+        !VncViewer.os.startsWith("mac os x")) {
       boolean isExtEvent = false;
       try {
         isExtEvent = processExtInputEvent(type);
@@ -517,6 +593,136 @@ public class Viewport extends JFrame {
       }
     }
     return retval;
+  }
+
+  // At the moment, these two methods are used only by the Mac TurboVNC Helper.
+  void handleTabletProximityEvent(final boolean enteringProximity,
+                                  final int pointingDeviceType) {
+    if (devices == null) return;
+
+    try {
+      SwingUtilities.invokeLater(
+        new Runnable() {
+          public void run() {
+            if (enteringProximity) {
+              switch (pointingDeviceType) {
+              case 1:  // pen
+                lastEvent.deviceID = 0;  // Stylus
+                break;
+              case 3:  // eraser
+                lastEvent.deviceID = 1;  // Eraser
+                break;
+              }
+            } else
+              lastEvent.deviceID = -1;
+          }
+        });
+    } catch (Exception e) {
+      vlog.error("SwingUtilities.invokeLater() failed: " + e.getMessage());
+    }
+  }
+
+  void handleTabletEvent(final int type, final double x, final double y,
+                         final float pressure, final float tiltX,
+                         final float tiltY) {
+    if (devices == null) return;
+
+    final int NSLeftMouseDown = 1;
+    final int NSLeftMouseUp = 2;
+    final int NSRightMouseDown = 3;
+    final int NSRightMouseUp = 4;
+    final int NSMouseMoved = 5;
+    final int NSLeftMouseDragged = 6;
+    final int NSRightMouseDragged = 7;
+    final int NSOtherMouseDown = 25;
+    final int NSOtherMouseUp = 26;
+    final int NSOtherMouseDragged = 27;
+
+    try {
+      SwingUtilities.invokeLater(
+        new Runnable() {
+          public void run() {
+            if (lastEvent.deviceID < 0)
+              // No prior proximity event was received, so we don't know which
+              // tablet device is generating this event.  Punt.
+              return;
+
+            Dimension winSize = sp.getSize();
+            java.awt.Point spOffset = sp.getViewport().getViewPosition();
+            ExtInputDevice dev = null;
+            for (ExtInputDevice d : devices) {
+              if (lastEvent.deviceID == d.id && d.remoteID != 0)
+                dev = d;
+            }
+            if (dev == null)
+              return;
+
+            if (type == NSLeftMouseDown || type == NSRightMouseDown ||
+                type == NSOtherMouseDown)
+              lastEvent.type = giiTypes.giiButtonPress;
+            else if (type == NSLeftMouseUp || type == NSRightMouseUp ||
+                     type == NSOtherMouseUp)
+              lastEvent.type = giiTypes.giiButtonRelease;
+            else
+              lastEvent.type = giiTypes.giiValuatorAbsolute;
+
+            lastEvent.buttonNumber = 0;
+            if (type == NSLeftMouseDown || type == NSLeftMouseUp)
+              lastEvent.buttonNumber = 1;
+            else if (type == NSRightMouseDown || type == NSRightMouseUp)
+              lastEvent.buttonNumber = 2;
+            else if (type == NSOtherMouseDown || type == NSOtherMouseUp)
+              lastEvent.buttonNumber = 3;
+
+            lastEvent.firstValuator = 0;
+            lastEvent.numValuators = 5;
+
+            double xtmp = (double)x;
+            if (dx > 0)
+              xtmp -= (double)dx;
+            xtmp += spOffset.x;
+            if (cc.cp.width != cc.desktop.scaledWidth) {
+              xtmp = (cc.desktop.scaleWidthRatio == 1.00) ? xtmp :
+                      xtmp / cc.desktop.scaleWidthRatio;
+            }
+            ExtInputDevice.Valuator v =
+              (ExtInputDevice.Valuator)dev.valuators.get(0);
+            lastEvent.valuators[0] = (int)(xtmp / (double)(cc.cp.width - 1) *
+                                     (double)(v.rangeMax - v.rangeMin) +
+                                     (double)v.rangeMin + 0.5);
+            if (lastEvent.valuators[0] > v.rangeMax)
+              lastEvent.valuators[0] = v.rangeMax;
+            else if (lastEvent.valuators[0] < v.rangeMin)
+              lastEvent.valuators[0] = v.rangeMin;
+
+            double ytmp = (double)sp.getSize().height - y - 1.0;
+            if (dy > 0)
+              ytmp -= (double)dy;
+            ytmp += spOffset.y;
+            if (cc.cp.height != cc.desktop.scaledHeight) {
+              ytmp = (cc.desktop.scaleHeightRatio == 1.00) ? ytmp :
+                     ytmp / cc.desktop.scaleHeightRatio;
+            }
+            v = (ExtInputDevice.Valuator)dev.valuators.get(1);
+            lastEvent.valuators[1] = (int)(ytmp / (double)(cc.cp.height - 1) *
+                                     (double)(v.rangeMax - v.rangeMin) +
+                                     (double)v.rangeMin + 0.5);
+            if (lastEvent.valuators[1] > v.rangeMax)
+              lastEvent.valuators[1] = v.rangeMax;
+            else if (lastEvent.valuators[1] < v.rangeMin)
+              lastEvent.valuators[1] = v.rangeMin;
+
+            lastEvent.valuators[2] = (int)(pressure * 65536.0 + 0.5);
+            lastEvent.valuators[3] = (int)(tiltX * 63.0 + 0.5);
+            lastEvent.valuators[4] = (int)(tiltY * 63.0 + 0.5);
+
+            lastEvent.print();
+            cc.writer().writeGIIEvent(dev, lastEvent);
+          }
+        });
+    } catch (Exception e) {
+      vlog.error("SwingUtilities.invokeLater() failed: " + e.getMessage());
+    }
   }
 
   private native void x11FullScreen(boolean on);